testing/test_io.py: refactor exception tests
The earlier implementation marked the tests with `pytest.mark.xfail`
(expected failure), which implies it is an external issue being worked
around, when it is actually intended behaviour.  The current
implementation actually separates the passing and failing tests, and
refactors the failures as tests that check if the correct exception is
being raised.

NB: on a more pratical note, this semantic difference is also how
certain IDEs (like vscode) interprets "expected to fail" unit tests,
and skip running them.