testing: sync with registry changes
registry now has a minimal set of columns, this led to many tests
failing.  had to rename many columns in data and index files.  several
files also had to be renamed for consistency.  this also introduced
several warnings for columns not in the registry, this is unavoidable
and can only be resolved after equivalent columns are added back to
the registry.